WIP: middle of extracting out types
